requirements.txt.txt
aiogram
python-dotenv
openai
aiohttp